Обновил задание 17 и 18

+++ src/ru/croc/task18/marketplace/ProductDao.java
@@ -4,40 +4,41 @@ import ru.croc.task18.marketplace.Order;
 import ru.croc.task18.marketplace.Product;
 
 import java.sql.*;
+import java.util.ArrayList;
 import java.util.List;
 
 public class ProductDao {
 
     public Product findProduct(String productCode){
-        try(Connection connection = DriverManager.getConnection("jdbc:h2:~/marketplace", "sa", "")) {
-            PreparedStatement getProduct = connection.prepareStatement("select * from `product` where vendorCode = ?");
+        try(Connection connection = DriverManager.getConnection("jdbc:h2:mem:~/test", "sa", "")) {
+            PreparedStatement getProduct = connection.prepareStatement("select * from `product` where vendorCode =?");
             getProduct.setString(1, productCode);
             ResultSet resultSet = getProduct.executeQuery();
-            if(!resultSet.next()) {
-                throw new Exception("Product not found!");
+            if(resultSet.next()) {
+                return new Product(resultSet.getString("vendorCode"),
+                        resultSet.getString("name"),
+                        resultSet.getInt("price"));
             }
-            return new Product(resultSet.getString("vendorCode"),
-                    resultSet.getString("name"),
-                    resultSet.getInt("price"));
+            return null;
         }
-        catch (Exception e) {
+        catch (SQLException e) {
             return null;
         }
     }
 
     public Product createProduct(Product product) throws SQLException {
-       try(Connection connection = DriverManager.getConnection("jdbc:h2:~/marketplace", "sa", "")){
+       try(Connection connection = DriverManager.getConnection("jdbc:h2:mem:~/test", "sa", "")){
            String vendorCode = product.getVendorCode();
            String name = product.getName();
            int price = product.getPrice();
 
-           String productRow = " values ('" + vendorCode + "', '" + name + "', '" + price + "')";
-           PreparedStatement productStatement = connection.prepareStatement("select * from `product` where vendorCode = ?");
-           productStatement.setString(1, vendorCode);
-           ResultSet productSet = productStatement.executeQuery();
-           if(!productSet.next()){
-               productStatement.execute("insert into `product` (vendorCode, name, price)" + productRow);
+           if(findProduct(product.getVendorCode()) != null) {
+               return null;
            }
+           String productRow = " values ('" + vendorCode + "', '" + name + "', " + price + ")";
+           Statement productStatement = connection.createStatement();
+           productStatement.execute("insert into `product` (vendorCode, name, price)" + productRow);
+
            return findProduct(vendorCode);
        }
         catch (SQLException e){
@@ -46,13 +47,21 @@ public class ProductDao {
     }
 
     public Product updateProduct(Product product) {
-        try(Connection connection = DriverManager.getConnection("jdbc:h2:~/marketplace", "sa", "")) {
-            Product updatedProduct = findProduct(product.getVendorCode());
+        try(Connection connection = DriverManager.getConnection("jdbc:h2:mem:~/test", "sa", "")) {
+            if(findProduct(product.getVendorCode()) == null) {
+                return null;
+            }
 
-            deleteProduct(updatedProduct.getVendorCode());
-            updatedProduct.setName(product.getName());
-            updatedProduct.setPrice(product.getPrice());
-            return createProduct(updatedProduct);
+            PreparedStatement productStatement = connection.prepareStatement("select * from `product` where vendorCode = ?");
+            productStatement.setString(1, product.getVendorCode());
+            ResultSet resultSet = productStatement.executeQuery();
+            resultSet.next();
+            int productId = resultSet.getInt("id");
+
+            Statement statement = connection.createStatement();
+            statement.execute("update product set vendorCode = '" + product.getVendorCode() + "',name = '" +
+                    product.getName() + "', price = " + product.getPrice() + " where id =" + productId);
+            return findProduct(product.getVendorCode());
         }
         catch (SQLException e) {
             return null;
@@ -60,21 +69,24 @@ public class ProductDao {
     }
 
     public void deleteProduct(String productCode) {
-        try(Connection connection = DriverManager.getConnection("jdbc:h2:~/marketplace", "sa", "")) {
-            Product product = findProduct(productCode);
+        try(Connection connection = DriverManager.getConnection("jdbc:h2:mem:~/test", "sa", "")) {
+            if(findProduct(productCode) == null) {
+                return;
+            }
 
-            PreparedStatement productStatement = connection.prepareStatement("select * from `product` where vendorCode = ?");
+            PreparedStatement productStatement = connection.prepareStatement("select * from `product` where vendorCode =?");
             productStatement.setString(1, productCode);
             ResultSet resultSet = productStatement.executeQuery();
-            int id = resultSet.getInt(1);
+            resultSet.next();
+            int productId = resultSet.getInt(1);
 
             productStatement = connection.prepareStatement("delete from `order` where product_id=?");
-            productStatement.setInt(1, id);
-            productStatement.executeQuery();
+            productStatement.setInt(1, productId);
+            productStatement.execute();
 
             productStatement = connection.prepareStatement("delete from `product` where vendorCode=?");
             productStatement.setString(1, productCode);
-            productStatement.executeQuery();
+            productStatement.execute();
         }
         catch (SQLException e) {
             e.printStackTrace();
@@ -82,10 +94,41 @@ public class ProductDao {
     }
 
     public Order createOrder(String userLogin, List<Product> products) {
+        try(Connection connection = DriverManager.getConnection("jdbc:h2:mem:~/test", "sa", "")){
 
+            PreparedStatement userStatement = connection.prepareStatement("select * from `user` where login =?");
+            userStatement.setString(1, userLogin);
+            ResultSet userSet = userStatement.executeQuery();
+            userSet.next();
+            int userId = userSet.getInt(1);
 
+            PreparedStatement orderStatement = connection.prepareStatement("select * from `order`");
+            ResultSet orderSet = orderStatement.executeQuery();
+            orderSet.last();
+            int orderId = orderSet.getInt(1) + 1;
 
-        return new Order();
+            List<Integer> userIdList = new ArrayList<>();
+            List<Integer> productIdList = new ArrayList<>();
+            for(Product product : products) {
+                String vendorCode = product.getVendorCode();
+
+                PreparedStatement productStatement = connection.prepareStatement("select * from `product` where vendorCode =?");
+                productStatement.setString(1, vendorCode);
+                ResultSet productSet = productStatement.executeQuery();
+                productSet.next();
+                int productId = productSet.getInt(1);
+
+                Statement statement = connection.createStatement();
+                String orderRow = " values (" + orderId + ", " + userId + ", " + productId +")";
+                statement.execute("insert into `order` (id, user_id, product_id)" + orderRow);
+                userIdList.add(userId);
+                productIdList.add(productId);
+            }
+            return new Order(userIdList, productIdList);
+        }
+        catch (SQLException e){
+            return null;
+        }
     }
 
 }
